Fixed &delete again

+++ commands/delete.js
@@ -1,11 +1,16 @@
 const tweetindex = require("../schema/tweetindex.js");
 const longoman = require("../helpers.js");
+const voter = require("../tweetumsVoter");
 
-module.exports.command = (message, args) => {
-    //Deletes all tweets, maybe.
-    if (longoman.isAdmin(message)) {
-        tweetindex.deleteOne({
-            sort: args[0],
+module.exports.command = async (message, args) => {
+    if(parseInt(args[0]) && longoman.isAdmin(message))
+    {
+        const deleted = await tweetindex.deleteOne({
+            sort: parseInt(args[0]),
         });
+        voter.fixSort();
+    }
+    else{
+        message.reply("unable to comply, partner.")
     }
 };
